patches, config changes, and documentation changes to get service working
There is an annoying problem with Python requests which causes the
whole service process to crash with a segmentation-fault if you use an
HTTPS connection back to ERMrest but have SE-Linux blocking the
`memexec` capability, such as a typical Fedora system with a sandboxed
Apache HTTPD. Setting the SE-Linux toggle `httpd_execmem` to `on`
will bypass this problem, but weakens the security posture of your
whole HTTPD and all its services.

Another alternative is to use plain HTTP. These changes to ERMresolve
add a top-level configuration document with an optional `"server_url"`
as well as per-target `"server_url"`. These allow an admin to
customize the resolver's communications with ERMrest. They also impact
the resulting URL provided to the client when resolution succeeds.
The configuration model is:

1. Service-wide `server_url` is set to `http://fqdn` by default.
2. Top-level config `server_url` overrides service-wide default.
3. Per-target config `server_url` overrides service-wide setting for
   that target only.