enhance default configuration behavior and revise documentation
The configuration file is completely optional. A missing file is the
same as a file containing just `{}` which will be augmented with
default field values to produce a multi-tenant resolver using
anonymous `/entity_rid/` requests for resolution URLs with embedded
catalog identifiers, e.g. `7/xyz`.

If a configuration file is present, but not usable, an internal error
is raised and logged for the administrator to resolve. This includes
basic JSON syntax errors as well as access/permissions problems for
the configuration file.  It will only use the default when it appears
the administrator is not trying to provide a configuration.